requirements.txt
streamlit==1.28.2
Pillow==10.1.0
google-generativeai==0.3.2
pytest==7.4.3